Fix the line chart tooltip
Line chart took a bit different approach in terms of tooltip: because we
showed large content, we wanted it to look reasonable in whichever
orientation appearing above the sidebar using a technical called portal.

The way we portaled previously in Polymer1 no longer works in Polymer2
especially because how real shadow CSS works. Unlike ShadyCSS where it
populated global style tag that is accessible anywhere in the DOM tree,
now, the styles are properly encapsulated inside a component. As a
result, styling for the vz-chart-tooltip became quite difficult.

With this change, to get around the styling issue, the vz-chart-tooltip
now take the name of the custom element and instantiate the element
(of course, with its style defined inside) and attached it to the body
of the document. When doing it, the vz-chart-tooltip style it a bit.