Fix issue where we werent initializing knn properly (#2200)
While debugging the standalone projector app, we discovered a bug caused
by the divergence between the umap-js and the `umap.d.ts` typings that
we maintain in this repo. This bug caused an error when initializing
very small datasets. This PR fixes the incorrect usage of
`umap.initializeFit` and updates the types to match the library.