build: fix lint errors
Summary:
Changes in #2267 displeased the Google linter: `list(my_depset)` is
banned because it’s not obvious that this forces iteration, so we switch
to the far clearer `.to_list()` everywhere.

The venerable linter further demands that docstrings be added, even for
drive-by fixes.

Test Plan:
A test import succeeds; all Google-internal tests pass.

wchargin-branch: import-lint