Polymer 2: fixed the Plottable interaction (#2249)
Plottable interactions are broken with the WebComponents due to its
event delegation and relative position computation -- their DOM
traversal was not compatible with the WebComponents. More information at
https://github.com/palantir/plottable/issues/3350.

Because of visibility (private and protected), it was not viable to
subclass the interactions to fix the issue so we have decided to monkey
patch the broken functions in JavaScript. In farther future, we may want
to use raw DOM events to handle the interactions instead of using the
Plottable abstraction.