core: always gate data selector behind query param
Summary:
We’d like to serve dashboard registration metadata from the backend, as
will be the case for dynamically loaded plugins. This is straightforward
except for the scalar plugin’s `useDataSelector` bit, which depends on
the query parameter. This patch changes moves the query parameter gating
from the scalar plugin to the core frontend.

Test Plan:
Verify that the following behavior is the same before and after this
change: (a) when the URL has no query parameters, neither the scalars
dashboard nor the images dashboard shows a data selector; and (b) when
the URL has `?EnableDataSelector`, the scalars dashboard has a data
selector and no run selector, while the images dashboard has a run
selector and no data selector.

wchargin-branch: top-level-data-selector-flag