mesh: fix steps slider
Summary:
The steps slider was supposed to be capped at the number of steps, but
in fact was always capped at 100 due to a typo in the Polymer binding.
Fixes #2270.

Test Plan:
Check out #2265 and run the mesh demo to generate test data. Then,
launch TensorBoard with the mesh data and observe that the slider now
goes from 0 to 9 instead of 0 to 100, and that moving along the entire
range of the slider is now useful.

wchargin-branch: mesh-fix-steps