build: upgrade `rules_closure` to HEAD
Summary:
This pulls in bazelbuild/rules_closure#383 to fix #2297, and integrates
the same fix into our own `Vulcanize.java` for consistency.

This diff includes a jscompiler upgrade from `v20190325` to `v20190528`.

Test Plan:
The `//tensorboard` target still builds and appears to run fine, with
cursory inspection.

Changing the `href` attribute of the help button in `tf-tensorboard` to
`mailto:xxx@yyy.com` now builds and works properly; prior to this
commit, that failed to build, as described in #2297.

wchargin-branch: rules-closure-mailto-fix