DO NOT SUBMIT: iframe loading proof of concept
Summary:
We create a same-origin iframe and add to it a `<script>` element that
dynamically imports the plugin entry point and invokes its `render`er.

For now, the plugin frame has no way to communicate with the host. In
the future, we’d like to be able to (e.g.) send “reload” signals from
host to plugin, and enable the plugin to query the host for global state
like the run selection.

Test Plan:
![Screenshot of hacked up scalars plugin](https://user-images.githubusercontent.com/4317806/58985670-c1531a00-8790-11e9-9b10-3fbe293c3c2d.png)

wchargin-source: c0e843435b5ea0ac3bf8740fc3f6eaa2a28f12dc
wchargin-branch: iframe-loading