Set base URI for plugin iframe
Plugins are developed with a relative to its package in mind. Without
setting the baseURI, the iframe has base path of
http[s]://[TBPATH]/[pathPrefix] and all relative paths that the plugin
uses require 'data/plugin/[name]/' prefix which can be unnatural. This
change sets the base URI of the plugin iframe to
http[s]://[TBPATH]/[pathPrefix]/data/plugin/[name]/.