Fixed incorrect runs-selector rendering
Nature of the bug:
Sidebar tf-runs-selector is a bit trickier in term of layout: its height
depends both on number/size of the content (# of runs) and size of the
viewport (show scrollbar if it does not fit in the sidebar).

To confine the size of the tf-runs-selector w.r.t the height of its
container (whose height is not defined by content but its sibling sizes
which are computed by respective descendents), we need to make each flex
container `overflow: hidden`. `overflow: hidden` makes child and
child-flex-container compute height w.r.t, the flex-container with
`overflow: hidden`.

When run numbers are low, we make tf-runs-selector's height computed
w.r.t. descendent heights by not putting `height: 100%`.

Confirmed the change both on Chrome 75 and Firefox 66.