build: add tool to create deterministic tarballs
Summary:
We plan to promote TensorBoard’s Pip packages to Bazel targets, but the
filenames of these targets can’t actually be statically determined: they
must include the TensorBoard version number, which is defined in a
Python module. Instead, we’ll build a tarball containing the two wheels.
Building tarballs deterministically with system tools is famously
troublesome (you can do it with GNU `tar` and a lot of flags, but not
portably), and it seems that the standard approach is to use Python’s
built-in `tarfile` library.

Implementation based off of <https://github.com/servo/servo/pull/12244>,
but with changes to make it Python 3-compatible, to use our code style,
and to clean up some minor issues and fix some TODOs from the original.
The original source is Apache-2 licensed.

Test Plan:
End-to-end tests included. Replacing the body of `_run_tool` with

```python
    return subprocess.check_output([
        "/bin/sh",
        "-c",
        'x="$(readlink -f "$1")" && cd "$2" && tar czf "$x" .',
        "unused",
    ] + args)
```

causes the `test_invariant_under_mtime` test case to fail for the right
reason.

wchargin-branch: deterministic-tgz