build: support `--incompatible_no_support_tools_in_action_inputs`
Summary:
One of many changes needed for Bazel 0.27.0; see #2355.

As mentioned in a comment, the fact that `collect_runfiles` continues to
be required surprises me: the `_tsc` `sh_binary` clearly declares `data`
dependencies on both Node and `tsc.js`, but those are apparently not
being included in the runfiles even when `_tsc` is passed in `tools`.

Test Plan:
The following command now works on Bazel 0.27.0:

```
bazel run \
  --incompatible_depset_is_not_iterable=true \
  --incompatible_disable_deprecated_attr_params=true \
  --incompatible_new_actions_api=true \
  --incompatible_no_support_tools_in_action_inputs=true \
  --incompatible_use_python_toolchains=false \
  //tensorboard -- --logdir ~/tensorboard_data
```

With the same `--incompatible_*` flags, all targets build and all tests
pass.

wchargin-branch: bazel-no-support-tools-in-action-inputs