diagnose: wrap diagnostics in `<details>` element (#2367)
Summary:
The output of `pip freeze` in particular can be multiple pages long. We
do want this output, but we don’t want to scroll through it.

Test Plan:
Run the diagnosis script and copy the output into a GitHub comment box
to preview the rendered Markdown. Note that it’s now compact—

![screenshot of rendered diagnostics output with `<details>` element][1]

—and that you can click the “Diagnostics output” to expand the full
contents.

[1]: https://user-images.githubusercontent.com/4317806/59795069-465d2980-928f-11e9-893c-d8c62d99d592.png

wchargin-branch: diagnostics-details