tf-runs-selector: make it take up parent's width. (#2366)
Especially when name of the run is short and size of the window is large,
tf-runs-selector was not taking up the size of the parent width. This bug
was quite subtle for our demo/test data because data location string was
long enough to make the tf-runs-selector take up wider width. By manually
removing the data location string (length of run names do not matter), we
were able to reproduce the bug.

This change makes tf-runs-selector take up the width of the parent.
Because we use position relative-absolute trick, we have to consider the
size of the parent's (or relative container) padding.

Fixes #2358.