build: add `:extract_pip_package` for convenience
Summary:
For ad hoc testing, one frequently wants to install a Pip package after
buliding it. A new script extracts the wheels and prints the path to the
enclosing directory (for `xdg-open` to drag-and-drop into Colab) and to
the wheels themselves (to copy-and-paste as a `pip install` argument).

Test Plan:
Run the new binary:

```
$ bazel run //tensorboard/pip_package:extract_pip_package 2>/dev/null
/tmp/tmp.caSTM9Sa7J
/tmp/tmp.caSTM9Sa7J/tensorboard-1.15.0a0-py2-none-any.whl
/tmp/tmp.caSTM9Sa7J/tensorboard-1.15.0a0-py3-none-any.whl
```

wchargin-branch: extract-pip-package