build: removed bin/genfile prefix in declare_file
Commit 4ba5f51b6d3beb0746ce21d355cef29d918d021d changed ctx.new_file to
ctx.actions.declare_file which is the new way of declaring a file except
the new API behaves a bit differently than the old one.

new_file, as the first argument, took the root directory [1] and
previous to the commit, we passed the bazel genfile path. However,
declare_file no longer take the root path [2] (it generates a new root
path based on whether the rule is genrule or not [3]) and appended the
first argument to the root it generates.

This made TensorBoard have paths like below:
- bazel-out/k8-fastbuild/bin/t/c/tf_backend/bazel-out/k8-fastbuild/bin/backend.js
- bazel-out/k8-fastbuild/bin/t/c/bazel-out/k8-fastbuild/bin/index.pbtxt

[1]:
https://github.com/bazelbuild/bazel/blob/2a116372382aedbf8a39477e0ea1123903ab5479/src/main/java/com/google/devtools/build/lib/analysis/skylark/SkylarkRuleContext.java#L759-L763
[2]: https://github.com/bazelbuild/bazel/blob/1ace20a6bb7587ae427d5917ffe13fa71d8885e9/src/main/java/com/google/devtools/build/lib/analysis/skylark/SkylarkActionFactory.java#L133
[3]: https://github.com/bazelbuild/bazel/blob/2a116372382aedbf8a39477e0ea1123903ab5479/src/main/java/com/google/devtools/build/lib/analysis/RuleContext.java#L621-L623

This changes unblock our tomservo process.