build: removed bin/genfile prefix in declare_file (#2406)
Commit 4ba5f51 changed `ctx.new_file` to
`ctx.actions.declare_file` which is the new way of declaring a file except
the new API behaves a bit differently than the old one.

new_file, as the first argument, took the root directory 1 and
previous to the commit, we passed the bazel genfile path. However,
declare_file no longer take the root path 2 (it generates a new root
path based on whether the rule is genrule or not 3) and appended the
first argument to the root it generates.

This made TensorBoard have paths like below:
- bazel-out/k8-fastbuild/bin/t/c/tf_backend/bazel-out/k8-fastbuild/bin/backend.js
- bazel-out/k8-fastbuild/bin/t/c/bazel-out/k8-fastbuild/bin/index.pbtxt
- bazel-out/k8-fastbuild/bin/t/c/tf_backend/bazel-out/k8-fastbuild/bin/tf_backend-tsc.json

With the change, above look like below:
- bazel-out/k8-fastbuild/bin/t/c/tf_backend/backend.js
- bazel-out/k8-fastbuild/bin/t/c/index.pbtxt
- bazel-out/k8-fastbuild/bin/t/c/tf_backend/tf_backend-tsc.json

Looked for every declare_file instance and made sure we did not make the mistake.

Related: there were confusions around the difference between bazel-genfiles vs. bazel-bin.
As of 0.25, the Bazel team had merged two concept as there is no tangible difference and
only adds confusion.
https://groups.google.com/forum/#!msg/bazel-dev/5iq7n1soenA/xGM_M1NOCAAJ

This changes unblock our tomservo process.