core: introduce rules_nodejs and angular shim (#2472)
TensorBoard has been using custom Vulcanization to build the frontend
but we have plans to use better officially Bazel-team supported rule:
rules_nodejs. This adds the WORKSPACE dependency to the repo.

Also, to prove the correctness of the integration, this change also
Angular CLI generated (heavily trimmed down and modified) code with its
build rule.

The change also includes Angular example that is not connected to TensorBoard.