Allow plugins to declare wildcard routes. (#2476)
Currently, plugins must explicitly enumerate the routes that they serve via get_plugin_apps(). This is inconvenient in the case of plugins that provide a set of related routes, such as a directory full of static files, whose names need not be known in advance. This is the case, for instance, for dynamic plugins using Angular or React, where the build process produces a set of assets (including HTML, JS, CSS, source maps, etc.).

This PR allows a plugin to declare a wildcard route ending in /*, e.g. {'/foo/*': self._foo_wildcard_handler}. Any request beginning with the specified path prefix will be passed to the given handler.

Note the handler may do whatever it likes with these requests; they do not necessarily correspond to filesystem paths (as in the static assets use case). Thus this change also allows for RESTful apis, where object IDs and such are encoded as path elements. Any security concerns (e.g., regarding paths containing /../), are the responsibility of the handler.