Use `rel="noopener"` for `target="_blank"` links
Summary:
Links with `target="_blank"` open in a new tab, but by default that new
tab can control the page that opened it via the `window.opener` object.
Using `rel="noopener"` closes this security hole.

Test Plan:
No automated test for now; we tend to catch these in code review, so
this is a backfill. Manual verification:

```
$ git grep -o noopener '*.html' | wc -l
20
$ git grep -o _blank '*.html' | wc -l
20
```

wchargin-branch: rel-noopener