polymer2: fixed plugin dropdown UI issue (#2517)
Now that we have real ShadowDom, the input within the paper-input has to
be styled through mulitple layers of webcomponents, correctly. Normally,
we would be using ShadyCSS (CSS mixin did not make it to the spec) and a
mixin API exposed by appropriate WebComponents to achieve that but it is
not viable in this case. Specifically, paper-dropdown-menu allows
styling of paper-input using a mixin but in order to style input within
the paper-input, we need to use another mixin. ShadyCSS explicitly says
the nested mixin is not supported.

Instead of transforming with CSS, we decided to fix the issue by
pre-processing the tabName in JavaScript.

Also, we fixed the wrong CSS selector after paper-menu -> paper-listbox
change.

See:
https://github.com/webcomponents/polyfills/tree/master/packages/shadycss#known-issues