scalars: omit empty runs from `/tags` route (#2520)
Summary:
This is not a user-facing change, because the runs selector is populated
by the separate `/data/runs` endpoint. As we begin to consider generic
data providers (cf. #2490), it seems superfluous that a “list scalars”
call to a storage layer might be required to include _all_ run names in
the response even if only a few have relevant data and the rest will be
ignored, anyway.

Test Plan:
Tests updated. Manually confirmed that runs like `text_demo` continue to
appear in the scalar dashboard’s run selector even though they have no
scalar data.

wchargin-branch: scalars-omit-empty