Fix dynamic plugin bootstrap sequence in Firefox
Summary:
We initialize dynamic plugins by creating a new iframe and injecting
into it a `script` element with `import("path/to/index.js")`. This does
not work in Firefox, throwing what appears to be an internal error:
<https://bugzilla.mozilla.org/show_bug.cgi?id=1573262>

@stephanwlee had the idea to try wrapping the contents of the inner
frame’s script with `setTimeout(…, 0)`, which worked! It turns out that
we have to move the `baseURI` initialization into that deferred context,
too.

Fixes #2536.

Test Plan:
Launch TensorBoard and navigate to `/#projector`. Confirm that the
projector plugin loads properly in both Chrome 76 and Firefox 68. The
vanilla JS `example_basic` plugin also works in both now, too.

Co-authored-by: Stephan Lee <2547313+stephanwlee@users.noreply.github.com>
wchargin-branch: dynamic-plugin-firefox