data: add `--generic_data` flag and `TBContext` field (#2531)
Summary:
This patch adds a `--generic_data` flag that defaults to `auto`; in the
`auto` state, each plugin determines whether its usage of the generic
data APIs is stable enough. The flag can also be set to `true` for
testing and dogfooding, or to `false` as an escape hatch in case some
bugs make it into the plugins deemed stable. Setting the flag to `false`
also skips constructing the data provider object itself.

wchargin-branch: generic-data-flag