projector: add images to nearest neighbors list (#2543)
Summary:
In addition to showing the labels of nearest neighbors, the projector
also shows sprites associated with those points. Sometimes the label
data alone is only minimally useful (e.g., when all the labels are the
same), and the sprites provide much greater insight. The additional
detail due to the larger view is also particularly helpful on datasets
with higher-resolution sprites, like audio spectrograms.

The behavior is enabled by default, and can be disabled via a new “show
images” checkbox.

Addresses part of #2492. 

Screenshots:

  - <https://user-images.githubusercontent.com/4317806/63119171-570ae280-bf54-11e9-8e83-80e19701a60f.png>
  - <https://user-images.githubusercontent.com/6741720/62971619-4a657d80-bde0-11e9-84d2-f7b9f56b8bf6.png>

Test Plan:
Launch the projector on a dataset without images, and observe that the
behavior is the same as before: no images, no new checkbox, no error
messages. Then, switch to a dataset with images, and select a point.
Verify that the “show images” toggle works, that the images are properly
rendered, and that the images update when you select a new data point.