build: remove `//tensorboard:tf_contrib_ffmpeg`
Summary:
As of #2557, we’ve migrated all users off of `tf.contrib.ffmpeg`, so
this BUILD target can be removed.

Test Plan:
Running `git grep contrib.ffmpeg` yields no results, and the only
mentions of `ffmpeg` are in `beholder/video_writing.py` (which shells
out to `ffmpeg(1)` as a subprocess).

wchargin-branch: remove-ffmpeg