core: expose backend and storage as Polymer
This change contains a hack that should improve developer experience
by allowing Angular app to wrap Polymer code.

In order for the Angular app to be useful, we need to cross the
JSCompiler and rollup boundary where symbols are separtely mangled.
By exposing as property in the Polymer element, we can make the symbol
more stable.