scalar: fix tiny chart
This change updates `active` prop to a template of a
tf-category-paginated-view when the category pane is collapsed/expanded.
It also updates the `active` when a component is no longer mounted onto
the DOM.

The change does not contain any redraw logic but it magically works
because the DataLoaderBehavior, when active changes to active, fetches
data point for ones that are not in cache and then triggers rerender.
This flow of logic is indirect and thus can be made to be more direct
but that can cause too many "redraw" call which can be quite expensive
(scales poorly with # of runs).