core: remove unused `mode` on `/data/environment`
Summary:
The `data_location` field will soon be made pluggable on the data
provider, which will muddy the semantics of the `mode` field. Given that
it’s unused, we can just remove it.

Test Plan:
There are no other uses of the removed `getMode` method:

    $ git grep -ho 'getMode\w*' | sort -u
    getModelClass
    getModelHeaderClass
    getModelName
    getModelName_

The frontend still seems to work fine in both logdir and DB modes.

wchargin-branch: remove-environ-mode