data: expose `data_location` integration point
Summary:
Generic data providers can now control the “data location” string in the
TensorBoard UI, which appears at the bottom of the runs selector:

![Screenshot of the runs selector and its data location string][ss]

As this is only a cosmetic change, it’s provided as an optional method
on the `DataProvider` interface.

[ss]: https://user-images.githubusercontent.com/4317806/64572837-cf7b7e00-d31d-11e9-8fd6-d81d36774172.png

Test Plan:
TensorBoard still seems to work properly with `--generic_data` set to
`false`, `auto`, or `true`, in both logdir and DB modes. Conspicuously
changing the `MultiplexerDataProvider` to always `return "tada"` induces
the appropriate change in the frontend, so the wiring is hooked up.

wchargin-branch: generic-data-location
wchargin-source: 34124f25978fa827704b106283c03c00c653198c