core: remove unused `mode` on `/data/environment` (#2626)
Summary:
The `data_location` field will soon be made pluggable on the data
provider, which will muddy the semantics of the `mode` field. Given that
it’s unused, we can just remove it.

(The semantics were already a bit wonky when both `--logdir` and `--db`
were specified, as `data_location` would point to the logdir but `mode`
would be `db`. Removing it entirely fixes this problem.)

Test Plan:
There are no other uses of the removed `getMode` method:

    $ git grep -ho 'getMode\w*' | sort -u
    getModelClass
    getModelHeaderClass
    getModelName
    getModelName_

The frontend still seems to work fine in both logdir and DB modes.

wchargin-branch: remove-environ-mode