hparams: remove `Experiment` from public API docs
Summary:
This was an error; the `Experiment` symbol never existed in the final
version of the public APIs. (It existed in earlier drafts, but was
removed for the sake of simplicity.)

Thanks to @tfboyd for reporting this.

Test Plan:
Note that this now matches the invocation in the demo that can actually
be run:

https://github.com/tensorflow/tensorboard/blob/9335e0d3a271293919822aa8a689e9ef04cd6ff8/tensorboard/plugins/hparams/hparams_demo.py#L215-L216

wchargin-branch: hparams-no-experiment