build: make tensor_widget build with jscomp (#2657)
The name, tensor_widget_binary, is unique to rollup_bundle; could not
figure out how to disable such global export in the bundler yet. As
such, we need to explicitly export a module to a global variable.

The externs file was introduced so the exported names are not mangle and
can remain stable.