projector: visualize_embeddings API change
Previously, it used to take an FileWriter as an input so it can infer
logdir to write embedding to. Because summary v2 API deprecated an
ability to get the logdir, we now instead take the logdir as string.

To work with the older code, we make sure we can also take the v1
summary writer.