Define a separate function to handle slice keys
Each slice of the dataset is identified by a key. Currently, when
slicing by a feature, each value becomes a slice and therefore a key.
To facilitate how we handle with generating slices, and particularly to
allow us to group multiple values into the same key (buckets for
numerical features), we wrap the slice key generation in a separate
function.
This commit doesn't change any aggregating behaviour but makes way to
the next steps.