Make breakdownFeatureSelected_ use slice key
We iterate over the dataset using the slice key function to attribute a
key to each example. This avoids the usage of the `undefined` dummy
keys, since all cases will be naturally covered by seeing all examples.
Since we are now using a single function to handle slice keys throughout
the code, we can change the key attribution logics in a consistent way.