Aggregate numeric features in buckets when slicing
We create buckets for each aggregated feature, so that we can associate
each slice of that feature with an interval instead of a value.
The getSliceKey_ function now checks if the feature has bucket to be
aggregated and returns the slice key of the appropriate interval in the
positive case. It uses the value of the feature as key otherwise (just as
before).