Order intervals when slicing by buckets
When slicing by numeric features aggregated in buckets, displaying
the buckets in order is a particularly useful feature.
Alphabetical ordering is not enough for this purpose: we need to
parse the number in the interval and order it by its value.
Also, let's make this ordering the default one when aggregating.