Fix wrong info dialogs
When tapping on an info button, the first paper-dialog DOM in the button's grandparent was opening. This caused a problem when all elements had the same grandparent, so that clicking any one of them would open the first dialog.
I've tested the other cases and they are all working fine.