core: implement experimental API for runs.
With tb.plugin.lib.run, you can now fetch list of runs and subscribe to
changes.

Do note that this change includes changes to the BUILD where we no
longer use modules for the library. We will revisit this in near future
to make this more usuable for non-TB Polymer projects