vulc: jscomp-ignored script in topological order
For jscomp-ignored script, we took a special treatment to place it
before the first script tag we found. Not sure why it is done that way
but it does not seem to cause functional regression.

This change allows us to place a script tag before loading polymer
binary (which is jscomp-ignored) so we can enable the Polymer's
strictTemplatePolicy.