Easier WIT custom_predict_fn when using JSON inputs and other witwidget 1.4 fixes (#2716)
If data is provided to WIT as JSON (not example protos) and the user defines a custom prediction function, the examples passed to that function should be in that JSON format, not in tf.Example format. This makes it much easier to use custom predict functions.

Also fix indexing error during fairness thresholding and fix log level setting.