application: add middleware for experiment IDs (#2735)
Summary:
This commit teaches TensorBoard to recognize URLs like `/experiment/123`
as specifying experiment IDs. If these two path components are omitted,
the experiment ID is the empty string, for backward compatibility. The
experiment ID is intercepted by middleware and stripped from the URL
before the application handles it, so no existing logic needs to change.
The ID is added to the WSGI environment and may be extracted via a new
public function, `plugin_util.experiment_id`.

This commit does not yet replace the existing query parameter wiring, so
this is not a user-facing change (except that more URLs are now valid).

This functionality is compatible with `--path_prefix`.

Note that the `/data/plugins_listing` implementation does not need to
change: it correctly specifies experiment IDs in iframe loading
mechanisms because we already use the general-purpose `SCRIPT_NAME` key
specified by WSGI for this purpose.

Test Plan:
Unit tests and a small integration test in `:application_test` included.
For manual testing, modify `application.py`’s `_route_request` method to
print out the application-level path and experiment ID:

```python
    # at top of `_route_request`:
    from tensorboard import plugin_util
    eid = plugin_util.experiment_id(environ)
    logger.warn("%r [%r]" % (environ.get("PATH_INFO"), eid))
```

Then, check that the scalars plugin and the dynamically loaded projector
plugin both work, and that the paths and experiment IDs are as expected,
in the following configurations:

  - with `--path_prefix=/foo`:
      - at `/foo` (should redirect to add slash)
      - at `/foo/`
      - at `/foo/experiment/123` (should redirect to add slash)
      - at `/foo/experiment/123/`
  - with no path prefix:
      - at `/`
      - at `/experiment/123` (should redirect to add slash)
      - at `/experiment/123/`

wchargin-branch: eid-middleware