backend: fix `--path_prefix` validation (#2737)
Summary:
Specifying a path prefix without leading slash has always been
disallowed by the docs, but in TensorBoard 2.0.0 this just hits an
inscrutable internal error when starting the server:

```
ValueError: invalid literal for int() with base 10: '6006my_path_prefix'
```

As of #2733, this is at least a correctly reported error, but it still
has a stack trace. This commit makes it a normal, fast-failing usage
error.

The logic to strip trailing slashes was also duplicated, so `/foo` and
`/foo/` and `/foo//` were all the same but differed from `/foo///`. This
commit removes one copy of the logic and makes the remaining copy strip
all trailing slashes instead of just one.

Test Plan:
Run with `--path_prefix /foo`, and note that TensorBoard exits with a
single-line error. Run with `/foo`, `/foo/`, `/foo//`, and `/foo///`,
and note that all four now have the same behavior.

wchargin-branch: path-prefix-validation