Add ROC curves for sliced features in multiclass
By considering one binarized problem for each class (the class vs. all
the rest), we build one ROC curve for each class.
This was done for the sliced case first so that the grouped case could
be treated as a simple extention of this one.
The data is stored for each model in `inferenceStats_`, in the key
`allThresholds`.